Refactor:remove FaqReader & FaqStore
