+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/faq/core/FaqReader.java
@@ -1,4 +1,4 @@
-package kernel.jdon.moduleapi.domain.faq.application;
+package kernel.jdon.moduleapi.domain.faq.core;
 
 import java.util.List;
 
@@ -6,5 +6,6 @@ import kernel.jdon.moduledomain.faq.domain.Faq;
 
 public interface FaqReader {
 	Faq findById(Long faqId);
+
 	List<Faq> findAll();
 }
