+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/faq/core/FaqStore.java
@@ -1,8 +1,9 @@
-package kernel.jdon.moduleapi.domain.faq.application;
+package kernel.jdon.moduleapi.domain.faq.core;
 
 import kernel.jdon.moduledomain.faq.domain.Faq;
 
 public interface FaqStore {
 	Faq save(Faq initFaq);
+
 	void delete(Faq persistFaq);
 }
