+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/faq/infrastructure/FaqReaderImpl.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import kernel.jdon.moduledomain.faq.domain.Faq;
-import kernel.jdon.moduleapi.domain.faq.application.FaqReader;
+import kernel.jdon.moduleapi.domain.faq.core.FaqReader;
 import kernel.jdon.moduleapi.domain.faq.error.FaqErrorCode;
+import kernel.jdon.moduledomain.faq.domain.Faq;
 import lombok.RequiredArgsConstructor;
 
 @Component
